+++ bcn_rainfall_core/__init__.py
@@ -1,5 +1,5 @@
 from bcn_rainfall_core.rainfall import Rainfall
 
-__version__ = "1.0.4"
+__version__ = "1.0.5"
 
 __all__ = ["Rainfall"]